refactor: useResize 훅 추가
